move filter struct
